requirements.txt
fastapi
uvicorn
docker
pyyaml
pydantic
requests
